+++ portal/app/[locale]/hemi-earn/pool/[shareAddress]/_components/withdraw.tsx
@@ -18,7 +18,7 @@ import { percentToBps } from '../../../_utils/slippage'
 import { usePoolForm } from '../_context/poolFormContext'
 import { useAssetsToShares } from '../_hooks/useAssetsToShares'
 import { useMaxWithdrawableAsset } from '../_hooks/useMaxWithdrawableAsset'
-import { useSlippage } from '../_hooks/useSlippageBps'
+import { useSlippage } from '../_hooks/useSlippage'
 import { useUserShareValue } from '../_hooks/useUserShareValue'
 import { useWithdraw } from '../_hooks/useWithdraw'
 import { useWithdrawPreview } from '../_hooks/useWithdrawPreview'
